refactor: #108 - TabBarList를 컴포넌트로 분리
